src/cloud_functions/alerts/compute_engine_on/requirements.txt
google-auth>=2.0.0
google-api-python-client>=2.0.0
dotenv